requirements.txt
Flask
python-vlc
requests
six
